+++ src/System/Log.hs
@@ -1,3 +1,10 @@
+{-# LANGUAGE CPP #-}
+{-# LANGUAGE DeriveDataTypeable #-}
+
+#if __GLASGOW_HASKELL__ >= 702
+{-# LANGUAGE DeriveGeneric #-}
+#endif
+
 {- |
    Module     : System.Log
    Copyright  : Copyright (C) 2004-2011 John Goerzen
@@ -25,6 +32,12 @@ module System.Log(-- * Types
 
     where
 
+import Control.DeepSeq (NFData(rnf))
+import Data.Data (Data, Typeable)
+#if __GLASGOW_HASKELL__ >= 702
+import  GHC.Generics (Generic)
+#endif
+
 {- | Priorities are used to define how important a log message is.
 Users can filter log messages based on priorities.
 
@@ -33,7 +46,7 @@ definitions are given below, but you are free to interpret them however you
 like.  They are listed here in ascending importance order.
 -}
 
-data Priority = 
+data Priority =
             DEBUG                   -- ^ Debug messages
           | INFO                    -- ^ Information
           | NOTICE                  -- ^ Normal runtime conditions
@@ -42,9 +55,15 @@ data Priority =
           | CRITICAL                -- ^ Severe situations
           | ALERT                   -- ^ Take immediate action
           | EMERGENCY               -- ^ System is unusable
-                    deriving (Eq, Ord, Enum, Bounded, Show, Read)
+#if __GLASGOW_HASKELL__ >= 702
+          deriving (Eq, Ord, Enum, Bounded, Show, Read, Data, Typeable, Generic)
+#else
+          deriving (Eq, Ord, Enum, Bounded, Show, Read, Data, Typeable)
+#endif
+
+-- | @since 1.3.1.0
+instance NFData Priority where rnf = (`seq` ())
 
 {- | Internal type of log records -}
 
 type LogRecord = (Priority, String)
-